WIP: make openshift-deploy a sibling to openshift-router
