--- images/deployer/kube-deploy/Dockerfile
@@ -1,3 +1,0 @@
-FROM scratch
-COPY kube-deploy /kube-deploy
-CMD ["/kube-deploy"]
